Added basic google-auth and bootstrap
A new called python-social-auth is added. Right
now the basic login with google account works.
You need to create an application with google and
use its secret and access. For testing purpose I
am adding my keys and secret here. Please don't
misuse it.

Add an entry in your /etc/hosts file which should
look like:
127.0.0.1   iithalumni.com

The reason for doing the above thing is a redirect url
is needed after google login and it shouldn't be
localhost or 127.0.0.1

Check it out in action by visiting:
http://iithalumni.com:8000/auth/

For initial bootstrapping run
bash bootstrap.sh